Move old persisters files, remove namespacing
